requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
jinja2==3.1.2
pillow==10.0.0
numpy==1.25.2
aiofiles==23.2.1
tflite-runtime
tensorflow==2.13.0
numpy==1.24.3
opencv-python==4.8.0.74
opencv-contrib-python==4.8.0.74
